requirements.txt
fastapi
uvicorn
sqlalchemy
pillow
requests
